+++ src/app.rs
@@ -202,6 +202,12 @@ impl App {
                 current_value: current_value_countdown,
                 elapsed_value: elapsed_value_countdown,
                 app_time,
+                // target time format is in sync how footer shows its local time
+                target_time_format: if footer_toggle_app_time == Toggle::On {
+                    Some(app_time_format)
+                } else {
+                    None
+                },
                 with_decis,
                 app_tx: app_tx.clone(),
                 vim_motions,
@@ -309,6 +315,7 @@ impl App {
                             app.local_time.set_app_time_format(format);
                         }
                         app.footer.set_app_time_format(new_format);
+                        app.countdown.set_app_time_format(new_format);
                     }
                 }
                 KeyCode::Char(',') => {